# docs/swagger.yaml
basePath: /
definitions:
...
  dto.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  dto.LoginResponse:
    properties:
      email:
        type: string
      id:
        type: integer
      role:
        type: string
      token:
        type: string
    type: object
  dto.NewUser:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  dto.PaginationMeta:
    properties:
      next_page:
        type: string
      page:
        type: integer
      prev_page:
        type: string
      total_page:
        type: integer
    type: object
  dto.RegisterResponse:
    properties:
      email:
        type: string
      id:
        type: integer
      role:
        type: string
    type: object
  dto.Response:
    properties:
      data: {}
      error:
        type: string
      meta:
        $ref: '#/definitions/dto.PaginationMeta'
      msg:
        type: string
      success:
        type: boolean
    type: object
...
host: localhost:5000
info:
  contact: {}
  description: Back End using go with gin.
  title: Tickitz Back End
  version: "1.0"
paths:
  /admin:
    get:
      consumes:
      - application/json
      description: Get list of all movies for admin management
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: Get all movies
      tags:
      - admin
  /admin/movies:
    post:
      consumes:
      - multipart/form-data
      description: Create a new movie (Requires admin token)
      parameters:
      - description: Movie Title
        in: formData
        name: title
        required: true
        type: string
      - description: Movie Synopsis
        in: formData
        name: synopsis
        required: true
        type: string
      - description: Movie Duration
        in: formData
        name: duration
        required: true
        type: integer
      - description: Movie Release Date (YYYY-MM-DD)
        in: formData
        name: release_date
        required: true
        type: string
      - description: Director ID
        in: formData
        name: director_id
        type: integer
      - description: Poster Image
        in: formData
        name: poster
        type: file
      - description: Backdrop Image
        in: formData
        name: backdrop
        type: file
      - collectionFormat: csv
        description: Genre IDs
        in: formData
        items:
          type: integer
        name: genre_ids
        type: array
      - description: Popularity Score
        in: formData
        name: popularity_score
        type: number
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: Create a movie
      tags:
      - admin
  /admin/movies/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a movie from the database (Requires admin token)
      parameters:
      - description: Movie ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: Delete a movie
      tags:
      - admin
    patch:
      consumes:
      - multipart/form-data
      description: Update movie details (Requires admin token)
      parameters:
      - description: Movie ID
        in: path
        name: id
        required: true
        type: integer
      - description: Movie Title
        in: formData
        name: title
        type: string
      - description: Movie Synopsis
        in: formData
        name: synopsis
        type: string
      - description: Movie Duration
        in: formData
        name: duration
        type: integer
      - description: Movie Release Date (YYYY-MM-DD)
        in: formData
        name: release_date
        type: string
      - description: Director ID
        in: formData
        name: director_id
        type: integer
      - description: Poster Image
        in: formData
        name: poster
        type: file
      - description: Backdrop Image
        in: formData
        name: backdrop
        type: file
      - description: Popularity Score
        in: formData
        name: popularity_score
        type: number
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: Update a movie
      tags:
      - admin
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and return JWT token
      parameters:
      - description: Login Credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.LoginResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.Response'
      summary: User login
      tags:
      - auth
  /auth/logout:
    delete:
      consumes:
      - application/json
      description: Invalidate JWT token by removing it from whitelist
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: User logout
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Create a new user account
      parameters:
      - description: User Registration Body
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dto.NewUser'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.RegisterResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Register a new user
      tags:
      - auth
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and then your token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"